requirements.txt
beautifulsoup4==4.7.1
celery==4.2.1
cmarkgfm==0.4.2
Django==2.1.7
django-bootstrap3>=11.0.0
django_celery_beat==1.4.0
django-celery-results==1.0.4
django-elasticsearch-dsl==0.5.1
django-environ==0.4.5
django-extensions>=2.1.5
django-model-utils>=3.1.2
django-recaptcha2==1.4.1
ipython==7.2.0
jsonschema==2.6.0
psycopg2-binary==2.7.7
PyGithub==1.43.5
Pygments==2.3.1
python-magic==0.4.15
PyYAML==4.2b4
readme_renderer[md]==24.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
watchdog[watchmedo]==0.9.0
whitenoise==4.1.2
